backend/requirements.txt
fastapi[standard]
openai
python-dotenv
librosa
numpy
transformers